Speedup bash prompt generation
Generation of the git information within the bash prompt previously used
`git status`. While that worked, it behaved really slowly in large git
repos. It was taking around 1.2 seconds to generate the prompt, which
felt really painful.

After making a series of optimizations, it is now down to just 0.012 on
average. That's a 100x speedup.